Add DayView, TileContent, WeekView components
